Removal of lastBuilt() check processing of headers
Because apache workers will not always have the lastBuilt tmp file of when dev/build was last run then the logical check of lastBuilt() would be falsey and then the CSP headers would not be processed.

The SRIRecord table was also not retaining the 'ModifiedTime' value so it could not check if previous SRIRecord record values matched the current Integrities.